requirements.txt

apache-airflow
apache-airflow-providers-google
pandas
google-cloud-bigquery
